% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm-utils.R
\name{apply_calls}
\alias{apply_calls}
\title{Apply SVM calls to all partitions of data split}
\usage{
apply_calls(data, svmcalls, datasplit, indexes = NULL)
}
\arguments{
\item{svmcalls}{placeholder}

\item{datasplit}{placeholder}

\item{indexes}{placeholder}
}
\value{
placeholder
}
\description{
Apply SVM calls to all partitions of data split
}
\examples{
placeholder

}
