% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm-utils.R
\name{apply_calls_bad}
\alias{apply_calls_bad}
\title{Apply SVM calls to all partitions of data split}
\usage{
apply_calls_bad(data, svmcalls, datasplit)
}
\arguments{
\item{svmcalls}{placeholder}

\item{datasplit}{placeholder}

\item{indexes}{placeholder}
}
\value{
placeholder
}
\description{
Apply SVM calls to all partitions of data split
}
\examples{
placeholder

}
