% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ArgSpecsBinaryProb-class}
\alias{ArgSpecsBinaryProb-class}
\title{Binary classification specification (probabilistic)}
\description{
Concrete \link[=ArgSpecs-class]{ArgSpecs} subclass for probabilistic binary tasks; built by
\code{\link[=random_machines]{random_machines()}} with \code{task = "binary"}, \code{prob = TRUE}.
}
