% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ArgSpecsClassifHard-class}
\alias{ArgSpecsClassifHard-class}
\title{Non-probabilistic classification specification}
\description{
Virtual parent of \link[=ArgSpecsBinary-class]{ArgSpecsBinary} and \link[=ArgSpecsMultiClass-class]{ArgSpecsMultiClass}:
predictions are class factors and the weighting metric scores hard classes
(e.g. \code{accuracy}).
}
\keyword{internal}
