% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ArgSpecsMultiClassProb-class}
\alias{ArgSpecsMultiClassProb-class}
\title{Multiclass classification specification (probabilistic)}
\description{
Concrete \link[=ArgSpecs-class]{ArgSpecs} subclass for probabilistic multiclass tasks; built
by \code{\link[=random_machines]{random_machines()}} with \code{task = "multiclass"}, \code{prob = TRUE}.
}
