% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ArgSpecsReg-class}
\alias{ArgSpecsReg-class}
\title{Regression specification}
\description{
Concrete \link[=ArgSpecs-class]{ArgSpecs} subclass for regression tasks; built by
\code{\link[=random_machines]{random_machines()}} with \code{task = "regression"}.
}
