% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm-utils.R
\name{boot_fit_omega}
\alias{boot_fit_omega}
\title{Apply SVM calls to all partitions of data split}
\usage{
boot_fit_omega(
  data,
  svmcalls,
  datasplit,
  metric_function,
  weight_function,
  indexes = NULL
)
}
\arguments{
\item{svmcalls}{placeholder}

\item{datasplit}{placeholder}

\item{indexes}{placeholder}
}
\value{
placeholder

placeholder
}
\description{
Apply SVM calls to all partitions of data split
}
\examples{
placeholder

Apply SVM calls to all partitions of data split

placeholder

}
