% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{BootModels}
\alias{BootModels}
\title{Helper constructor for BootModels}
\usage{
BootModels(trainData, bootData, models, lambdas)
}
\arguments{
\item{bootData}{Result slot from BootSamples}

\item{models}{Result slot from KernelLambdas}

\item{lambdas}{Result slot from KernelModels}
}
\value{
BootModels S4 object
}
\description{
Helper constructor for BootModels
}
