% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BootSamples-class}
\alias{BootSamples-class}
\title{An S4 class representing a user profile}
\description{
An S4 class representing a user profile
}
\section{Slots}{

\describe{
\item{\code{trainData}}{Training data to have bootstrap samples generated}

\item{\code{bootFun}}{Bootstrap function passed into object creation}

\item{\code{bootArgs}}{Arguments passed to bootstrap function}

\item{\code{bootData}}{Bootstrap data stored after samples are generated}
}}

