% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight-utils.R
\name{brier_normalize}
\alias{brier_normalize}
\title{Brier normalizing}
\usage{
brier_normalize(x)
}
\arguments{
\item{x}{a metrics vector for assigning weights}
}
\value{
a vector of weights
}
\description{
Brier normalizing
}
\examples{
Placeholder
}
