% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{brier_score}
\alias{brier_score}
\title{Brier scoring - default measure for probabilistic accuracies}
\usage{
brier_score(x)
}
\arguments{
\item{x}{a metrics vector for assigning weights}
}
\value{
a vector of weights
}
\description{
Brier scoring - default measure for probabilistic accuracies
}
\examples{
Placeholder
}
