% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight-utils.R
\name{brier_weighter}
\alias{brier_weighter}
\title{Brier weighter}
\usage{
brier_weighter(x)
}
\arguments{
\item{x}{metrics vector for assigning weights}
}
\value{
vector of weights
}
\description{
Brier weighter
}
\examples{
Placeholder
}
