% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm-utils.R
\name{call_builder}
\alias{call_builder}
\title{Call building utility for kernlab}
\usage{
call_builder(specs)
}
\arguments{
\item{specs}{placeholder}
}
\value{
placeholder
}
\description{
Call building utility for kernlab
}
\examples{
placeholder

}
