% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_entropy.R
\name{class_entropy}
\alias{class_entropy}
\title{Shannon entropy for a vector of class probabilities}
\usage{
class_entropy(probs)
}
\arguments{
\item{probs}{Vector of probabilities}
}
\value{
a value for shannon entropy
}
\description{
Shannon entropy for a vector of class probabilities
}
