% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_partitions.R
\name{create_partitions}
\alias{create_partitions}
\title{Map numeric features to an integer partition id by cutting each numeric
feature into n_partitions bins and encoding the multi-dimensional bin as a
single index. Non-numeric columns are ignored.}
\usage{
create_partitions(df, n_partitions = 4)
}
\arguments{
\item{df}{data.frame that will be partitioned}

\item{n_partitions}{number of partitions}
}
\value{
a factor of partitions
}
\description{
Map numeric features to an integer partition id by cutting each numeric
feature into n_partitions bins and encoding the multi-dimensional bin as a
single index. Non-numeric columns are ignored.
}
