% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BootModels-methods.R,
%   R/BootSamples-methods.R, R/KernelProb-methods.R, R/RMPredictor-methods.R,
%   R/predict-methods.R
\name{displayInfo}
\alias{displayInfo}
\alias{displayInfo,UserProfile-method}
\title{Methods to display information}
\usage{
displayInfo(object, ...)

\S4method{displayInfo}{UserProfile}(object)

\S4method{displayInfo}{UserProfile}(object)

\S4method{displayInfo}{UserProfile}(object)

\S4method{displayInfo}{UserProfile}(object)

\S4method{displayInfo}{UserProfile}(object)
}
\arguments{
\item{object}{A description of the argument.}
}
\value{
What it returns.
}
\description{
Description.
}
