% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm-utils.R
\name{.callbuider}
\alias{.callbuider}
\title{Call building utility for kernlab}
\usage{
.callbuider(specs)
}
\arguments{
\item{specs}{placeholder}
}
\value{
placeholder
}
\description{
Call building utility for kernlab
}
\examples{
placeholder

}
