% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight-utils.R
\name{exp_normalize}
\alias{exp_normalize}
\title{Exponential normalizing}
\usage{
exp_normalize(x, beta)
}
\arguments{
\item{x}{a metrics vector for assigning weights}

\item{beta}{penalty hyperparameter}
}
\value{
a vector of weights
}
\description{
Exponential normalizing
}
\examples{
Placeholder
}
