% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{exp_score}
\alias{exp_score}
\title{Exponential scoring - default approach for regression KernelLambdas and BootOmegas calculation}
\usage{
exp_score(x, beta)
}
\arguments{
\item{x}{a metrics vector for assigning weights}

\item{beta}{penalty hyperparameter}
}
\value{
a vector of weights
}
\description{
Exponential scoring - default approach for regression KernelLambdas and BootOmegas calculation
}
\examples{
Placeholder
}
