% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FittedRM-class}
\alias{FittedRM-class}
\title{An S4 class representing a user profile}
\description{
An S4 class representing a user profile
}
\section{Slots}{

\describe{
\item{\code{omega}}{A numeric vector containing final model weights
An S4 class representing a fitted Random Machines model
This object will store the RM lifecycle progress}
}}

