% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{KernelData}
\alias{KernelData}
\title{Initial model training data}
\description{
training data splits and validation setequal
}
\section{Slots}{

\describe{
\item{\code{data}}{list of training data splits}

\item{\code{splitfun}}{data splitting function}
}}

