% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localized_sampling.R
\name{localized_sampling}
\alias{localized_sampling}
\title{Pending Title}
\usage{
localized_sampling(
  data,
  y_var,
  final_sample_size = NULL,
  final_sample_prop = NULL,
  n_partitions = 6,
  heterogeneous_prop = 0.8,
  return_idx = FALSE
)
}
\arguments{
\item{data}{data.frame that will be sampled}

\item{n_partitions}{Number of partitions}
}
\description{
Pending Title
}
