% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_enemy_sampling.R
\name{nearest_enemy_sampling}
\alias{nearest_enemy_sampling}
\title{Sample rows with probability proportional to an exponential transform of the}
\usage{
nearest_enemy_sampling(data, y_var, final_sample_size, alpha = 1)
}
\arguments{
\item{data}{data.frame with features and y}

\item{y_var}{name of class column}

\item{final_sample_size}{number of rows to sample}

\item{alpha}{temperature/slope for the exponential scaling}
}
\value{
a Nearest-Enemy sampled data.frame
}
\description{
Sample rows with probability proportional to an exponential transform of the
}
