% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{NumOrFactor-class}
\alias{NumOrFactor-class}
\title{Virtual Class containing arguments passed to the main RM function}
\description{
Virtual Class containing arguments passed to the main RM function
}
