% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm-utils.R
\name{omega_calc}
\alias{omega_calc}
\title{Bootstrap model weight}
\usage{
omega_calc(bootMetrics, omegaFunction)
}
\arguments{
\item{bootMetrics}{placeholder}

\item{omegaFunction}{placeholder}
}
\value{
placeholder
}
\description{
Bootstrap model weight
}
\examples{
placeholder
}
