% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RM_classifier.R
\name{predict.randomMachinesClassifierClass}
\alias{predict.randomMachinesClassifierClass}
\title{predict a class according to a Random-Machines trained model}
\usage{
\method{predict}{randomMachinesClassifierClass}(rmc_model, newdata, type = "class")
}
\arguments{
\item{rmc_model}{Random-Machines model that will be used to predict new data}

\item{newdata}{Test dataset containing new data}

\item{type}{A character informing the required answer
Available options are:
\itemize{
\item \code{"class"}: Will return classes;
\item \code{"probabilitiies"}: Will return probabilites.
}}
}
\value{
a vector with the predicted classes or the predicted probabilities
}
\description{
predict a class according to a Random-Machines trained model
}
