% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RM_classifier.R
\name{randomMachinesClassifier}
\alias{randomMachinesClassifier}
\title{Train and evaluate a Random-Machines model}
\usage{
randomMachinesClassifier(
  formula,
  data,
  B = 50,
  method = "original",
  kernels = list(kernlab::vanilladot(), kernlab::polydot(2), kernlab::rbfdot(1),
    kernlab::laplacedot(1)),
  loss_function = accuracy,
  C = 1
)
}
\arguments{
\item{formula}{the classes that will be predicted (must be factor) and the covariables that
will be used to predict it}

\item{data}{the dataset to be used}

\item{B}{how many bootstrap samples that will be used in Random-Machines training}

\item{method}{Wich method will be used to train the model (original, NE for Nearest-Enemy Sampling,
LS for localized sampling and WSVM for Weak-SVM)}

\item{kernels}{Wich kernels will be used}

\item{loss_function}{Define the loss function in the training}

\item{C}{SVM regularization parameter}
}
\value{
a randomMachinesClassifierClass with the trained model, metrics, etc.
}
\description{
Train and evaluate a Random-Machines model
}
\examples{
randomMachinesClassifier()
}
