% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-utils.R
\name{reverse_bs}
\alias{reverse_bs}
\title{reverse_bs}
\usage{
reverse_bs(bs_result, original_data)
}
\arguments{
\item{bs_result}{placeholder}

\item{original_data}{placeholder}
}
\value{
placeholder
}
\description{
reverse_bs
}
\examples{
placeholder

}
