% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-methods.R
\name{rm_specs}
\alias{rm_specs}
\title{Title}
\usage{
rm_specs(
  data = iris,
  formula = Species ~ .,
  task = "binary",
  prob = FALSE,
  implementation = "kernlab",
  kernels = c("rbf", "laplace", "polydot"),
  args = list(rbf = list(C = 1, epsilon = 0.1, kernel = kernlab::rbfdot(sigma = 1)),
    laplace = list(C = 1, epsilon = 0.01, kernel = kernlab::laplacedot(sigma = 1)),
    polydot = list(C = 1, epsilon = 0.01, kernel = kernlab::polydot(degree = 1, scale =
    1))),
  B = 100,
  lambdaMetric = yardstick::accuracy_vec,
  lambdaFunction = log_normalize,
  omegaMetric = yardstick::accuracy_vec,
  omegaFunction = default_weight_binary
)
}
\arguments{
\item{data}{data.frame}

\item{formula}{formula defining model fit}

\item{task}{task to be performed. Either 'binary', 'multiclass' or 'regression'}

\item{prob}{TRUE for probabilistic model; default: FALSE}

\item{implementation}{'kernlab' is the only available backend for now}

\item{kernels}{character identifiers for the kernels and parameter combinations listed under 'args'}

\item{args}{list with arguments to be passed to kernlab's 'ksvm' function}
}
\value{
placeholder
}
\description{
Title
}
\examples{
placeholder

}
