% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{RMSpecs}
\alias{RMSpecs}
\title{RMSpecs constructor helper function}
\usage{
RMSpecs(
  x = x,
  y = y,
  task = task,
  prob = prob,
  kernels = kernels,
  hyperparams = hyperparams,
  b = b,
  lambdaMetric = lambdaMetric,
  lambdaFunction = lambdaFunction,
  omegaMetric = omegaMetric,
  omegaFunction = omegaFunction
)
}
\arguments{
\item{x}{training dataset}

\item{y}{target variable}

\item{task}{binary, classification or multiclass}

\item{kernels}{list of kernel functions to be used within kernlab}

\item{b}{number of bootstrap samples to be used}

\item{lambdaMetric}{performance metric for lambda calculation}

\item{lambdaFunction}{function to determine kernel sampling probability calculation from lambda metrics}

\item{omegaMetric}{performance metric for omega calculation}

\item{omegaFunction}{function to determine weights for each b model's predictions in final prediction}
}
\value{
RMSpecs S4 object
}
\description{
RMSpecs constructor helper function
}
