% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_sampler.R
\name{sample_bootstrap}
\alias{sample_bootstrap}
\title{Create multiple bootstrap samples from a database, reparting it with training and test samples}
\usage{
sample_bootstrap(data, n = 100)
}
\arguments{
\item{data}{the database that will be used to create the bootstrap samples}

\item{n}{number of bootstrap samples that will be created}
}
\value{
a list
}
\description{
Create multiple bootstrap samples from a database, reparting it with training and test samples
}
