% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_exponential.R
\name{scale_exponential}
\alias{scale_exponential}
\title{softmax-like exponential scaling with temperature alpha}
\usage{
scale_exponential(x, alpha = 1)
}
\description{
softmax-like exponential scaling with temperature alpha
}
