% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasplit-utils.R
\name{simpleHoldout}
\alias{simpleHoldout}
\title{Create simple holdouts}
\usage{
simpleHoldout(df, p, y, balanced = TRUE)
}
\arguments{
\item{df}{data frame that will be splited}

\item{p}{proportion data that will be used to train a model}

\item{y}{name of target column}

\item{balanced}{makes the folds have the same proportion of classes (TRUE by default)}
}
\value{
a list with train and test elements for modeling
}
\description{
Create simple holdouts
}
\examples{
placeholder
}
