% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitFunctions-utils.R
\name{stratfiedKfold}
\alias{stratfiedKfold}
\title{Create stratified kfold samples}
\usage{
stratfiedKfold(df, K = 5, y, balanced = TRUE)
}
\arguments{
\item{df}{a data frame}

\item{K}{number of folds, must be a integer}

\item{y}{name of target column}

\item{balanced}{makes the folds have the same proportion of classes (TRUE by default)}
}
\value{
a matrix
}
\description{
Create stratified kfold samples
}
