% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_kfold.R
\name{stratified_kfold}
\alias{stratified_kfold}
\title{Create stratified kfold samples}
\usage{
stratified_kfold(df, K = 5, y)
}
\arguments{
\item{df}{a data frame}
}
\description{
Create stratified kfold samples
}
