% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_o2.R
\name{sys_o2}
\alias{sys_o2}
\title{Systematic Sampling Over Order (SyS-O2)}
\usage{
sys_o2(
  data,
  y_var,
  n_total = NULL,
  prop = NULL,
  order_by = NULL,
  decreasing = FALSE,
  within = c("random", "given"),
  seed = NULL,
  return_idx = FALSE
)
}
\arguments{
\item{data}{data.frame/tibble}

\item{y_var}{class column name (factor/character/integer)}

\item{n_total}{desired total sample size (optional if using "prop")}

\item{prop}{global proportion (0<prop<=1) of the total (optional if using \code{n_total})}

\item{order_by}{column name to determine within-class ordering (optional)}

\item{decreasing}{if TRUE, sort decreasing by "order_by"}

\item{within}{"random" (default) or "given" (keep input order)}

\item{seed}{optional random seed}

\item{return_idx}{if TRUE, return only row indices}
}
\description{
Systematic Sampling Over Order (SyS-O2)
}
