% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ArgSpecs-class}
\alias{ArgSpecs-class}
\title{Title}
\value{
placeholder
}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{data}}{data.frame.}

\item{\code{formula}}{formula.}

\item{\code{task}}{character.}

\item{\code{prob}}{logical.}

\item{\code{implementation}}{character.}

\item{\code{kernels}}{character.}

\item{\code{args}}{list.}

\item{\code{B}}{numeric.}

\item{\code{lambdaMetric}}{function.}

\item{\code{lambdaFunction}}{function.}

\item{\code{omegaMetric}}{function.}

\item{\code{omegaFunction}}{function.}
}}

