% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ArgSpecsMultiClass-class}
\alias{ArgSpecsMultiClass-class}
\title{Multiclass classification specification (hard)}
\description{
Concrete \link[=ArgSpecs-class]{ArgSpecs} subclass for hard multiclass tasks; built by
\code{\link[=random_machines]{random_machines()}} with \code{task = "multiclass"}, \code{prob = FALSE}.
}
