% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BootOmega-class}
\alias{BootOmega-class}
\title{An S4 class representing a user profile}
\description{
An S4 class representing a user profile
}
\section{Slots}{

\describe{
\item{\code{omegaMetrics}}{placeholder}

\item{\code{omegaFunction}}{placeholder}
}}

