% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{BootOmegas}
\alias{BootOmegas}
\title{BootOmegas constructor function}
\usage{
BootOmegas(specs = specs, bootData, svmcalls, lambdas)
}
\value{
placeholder
}
\description{
BootOmegas constructor function
}
