% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{BootSamples}
\alias{BootSamples}
\title{BootSamples}
\usage{
BootSamples(trainData, bootFun = simple_bs, bootArgs)
}
\arguments{
\item{trainData}{Training data to be passed into object construction}

\item{bootFun}{}

\item{bootArgs}{}
}
\value{
placeholder
}
\description{
BootSamples helper constructor
}
\examples{
placeholder
}
