% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{BootSamples}
\alias{BootSamples}
\title{BootSamples helper constructor}
\usage{
BootSamples(trainData, bootFun = simple_bs, bootArgs)
}
\arguments{
\item{trainData}{Training data to be passed into object construction}

\item{bootFun}{Bootstrap function to be applied}

\item{bootArgs}{Arguments to bootstrap function}
}
\value{
a BootSamples object
}
\description{
Runs \code{bootFun} with \code{bootArgs} and wraps the result in a \link{BootSamples} object.
}
\examples{
BootSamples(iris, simple_bs, list(indexes = seq_len(nrow(iris)), B = 10))
}
