% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{brier_weighter}
\alias{brier_weighter}
\title{Brier weights for BootOmega - default for probabilistic classification}
\usage{
brier_weighter(x)
}
\arguments{
\item{x}{metrics vector for assigning weights}
}
\value{
vector of weights
}
\description{
Brier weights for BootOmega - default for probabilistic classification
}
\examples{
Placeholder
}
