% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{buildCall}
\alias{buildCall}
\title{Method to build function calls}
\usage{
buildCall(object, ...)
}
\arguments{
\item{object}{A description of the argument.}
}
\value{
list of calls
}
\description{
Should operate differently based on "implementation" (kernlab/e1071)
}
