% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{displayInfo}
\alias{displayInfo}
\title{Methods to display information}
\usage{
displayInfo(object, ...)
}
\arguments{
\item{object}{A description of the argument.}
}
\value{
What it returns.
}
\description{
Description.
}
