% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FittedRM-class}
\alias{FittedRM-class}
\title{Title FittedRM}
\value{
placeholder
}
\description{
Title FittedRM
}
\section{Slots}{

\describe{
\item{\code{specs}}{ArgSpecs.}

\item{\code{lambdas}}{KernelLambdas.}

\item{\code{bs_samples}}{BootSamples.}

\item{\code{boot_models}}{BootModels.}

\item{\code{boot_omega}}{BootOmega.}
}}

\examples{
placeholder

}
