% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{KernelLambdas-class}
\alias{KernelLambdas-class}
\title{Internal S4 class for RM 1st stage representation}
\description{
Internal S4 class for RM 1st stage representation
}
\section{Slots}{

\describe{
\item{\code{models}}{List of SVMs trained.}

\item{\code{splits}}{Matrix representing data splits used in loss computation}

\item{\code{errors}}{Per-split errors}

\item{\code{kprobs}}{resulting probabilities}
}}

