% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{KernelLambdas}
\alias{KernelLambdas}
\title{An S4 class representing bootsrap data - works with rsample "bootstraps" function by default}
\usage{
KernelLambdas(models, splits, loss, probfun, lambdas)
}
\description{
An S4 class representing bootsrap data - works with rsample "bootstraps" function by default
}
