% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{KernelModels-class}
\alias{KernelModels-class}
\title{Trained models for each KernelSamples}
\description{
Models (one per kernel) fitted to specified splits (or no split)
}
\section{Slots}{

\describe{
\item{\code{models}}{list containing all trained models}

\item{\code{data}}{list of data splits each training process used}
}}

