% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{KernelSamples}
\alias{KernelSamples}
\title{Initial model training data}
\usage{
KernelSamples(splitfun, splitargs)
}
\arguments{
\item{splitfun}{splitfunction that will be used to split the data}

\item{splitargs}{arguments of the split function}
}
\value{
a KernelSamples object
}
\description{
training data splits and validation setequal
}
\section{Slots}{

\describe{
\item{\code{data}}{list of training data splits}

\item{\code{splitfun}}{data splitting function}
}}

\examples{
Placeholder
}
