% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm-utils.R
\name{lambda_calc}
\alias{lambda_calc}
\title{Kernel probabilities}
\usage{
lambda_calc(kernelMetrics, omegaFunction)
}
\arguments{
\item{kernelMetrics}{placeholder}

\item{omegaFunction}{placeholder}
}
\value{
placeholder
}
\description{
Kernel probabilities
}
\examples{
placeholder
}
