% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{lambdaCalc}
\alias{lambdaCalc}
\title{Lambda calculation}
\usage{
lambdaCalc(metrics)
}
\arguments{
\item{metrics}{numeric vector containing metrics for each kernel function}
}
\description{
Kernel function sampling probability
}
