% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{lambdaCalc}
\alias{lambdaCalc}
\title{Lambdas calculation - setting the generic allows for building different functions as methods for the KernelLambdas class
Ideally, a restricted number of methods would exist to operate on the class to calculate Lambdas}
\usage{
lambdaCalc(metrics)
}
\arguments{
\item{metrics}{numeric vector containing metrics for each kernel function}
}
\value{
a numeric vector of kernel selection probabilities (summing to 1)
}
\description{
Kernel function sampling probability
}
\keyword{internal}
