% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RM_classifier.R
\name{predict.randomMachinesClassifierClass}
\alias{predict.randomMachinesClassifierClass}
\title{predict a class according to a Random-Machines trained model}
\usage{
\method{predict}{randomMachinesClassifierClass}(rmc_model, newdata, type = "class")
}
\arguments{
\item{rmc_model}{Random-Machines trained model}

\item{newdata}{Test dataset containing new observations}

\item{type}{class}
}
\value{
a vector with the predicted classes
}
\description{
predict a class according to a Random-Machines trained model
}
