% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RandomMachines-class}
\alias{RandomMachines-class}
\title{Fitted RandomMachines ensemble}
\description{
Fitted RandomMachines ensemble
}
\section{Slots}{

\describe{
\item{\code{specs}}{ArgSpecs.}

\item{\code{kernelSamples}}{KernelSamples.}

\item{\code{kernelLambdas}}{KernelLambdas.}

\item{\code{bootSamples}}{BootSamples.}

\item{\code{bootOmegas}}{BootOmegas.}
}}

