% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RM_classifier.R
\name{randomMachinesClassifier}
\alias{randomMachinesClassifier}
\title{Train and evaluate a Random-Machines model. The training can be accelerated using alternative sample methods}
\usage{
randomMachinesClassifier(
  formula,
  data,
  B = 50,
  method = "original",
  kernels = list(kernlab::vanilladot(), kernlab::polydot(2), kernlab::rbfdot(1),
    kernlab::laplacedot(1)),
  loss_function = accuracy,
  C = 1,
  n_wm_models = 10,
  wm_sample_size_prop = 0.1,
  worst_prop = 0.8
)
}
\arguments{
\item{formula}{A formula specifying the target variable (must be factor) and the predictor variables used in classification.}

\item{data}{A data frame containing the training dataset}

\item{B}{Number of bootstrap samples that will be used in Random-Machines training}

\item{method}{Sampling strategy that will be used to train the model
Available options are:
\itemize{
\item \code{"original"}: standard Random-Machines training;
\item \code{"NE"}: Nearest-Enemy Sampling;
\item \code{"LS"}: Localized Sampling;
\item \code{"WSVM"}: Weak-SVM based sampling.
}}

\item{kernels}{List containing the kernels to be used during training}

\item{loss_function}{Define the loss function in the training}

\item{C}{Cost/regularization parameter of the SVM models.}

\item{n_wm_models}{Number of weak SVM models that will be trained in the WSVM method
to reduce data to probable support vector observations}

\item{wm_sample_size_prop}{Proportion of the dataset that will be used to train each weak SVM model}

\item{worst_prop}{Proportion of observations selected based on prediction entropy among weak SVM models.}
}
\value{
a randomMachinesClassifierClass with the trained model, metrics, etc.
}
\description{
Train and evaluate a Random-Machines model. The training can be accelerated using alternative sample methods
}
