% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{simple_bs}
\alias{simple_bs}
\title{Simple bootstrap resampling}
\usage{
simple_bs(indexes, B)
}
\arguments{
\item{indexes}{integer vector of row indices to resample}

\item{B}{number of bootstrap resamples to generate}
}
\value{
\code{list(train, test)}: \code{train} holds the resampled indices
(column per resample), \code{test} flags the out-of-bag rows.
}
\description{
Draws \code{B} bootstrap resamples of \code{indexes}, returning the in-sample
(\code{train}) and out-of-bag (\code{test}) memberships as two matrices, one column
per resample.
}
\examples{
simple_bs(indexes = 1:10, B = 5)

}
