% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{simple_bs}
\alias{simple_bs}
\title{simple_bs}
\usage{
simple_bs(indexes, B)
}
\arguments{
\item{indexes}{character or string containing either data.frame row names or indexes}

\item{B}{scalar representing number of bootstrap samples to generate}
}
\value{
List containing column-per-resample matrix for bootstrap samples and indicator matrix for OOB samples matrix
}
\description{
Bootstrap, generating in-sample and out-of-sample elements in two lists
}
\details{
Simple boostrap function
}
\examples{
simple_bs(indexes = 1:10, B = 5)

}
