% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{simple_bs}
\alias{simple_bs}
\title{simple_bs}
\usage{
simple_bs(indexes, B)
}
\arguments{
\item{[arg_name]}{\link{Data type and description of the input}}
}
\value{
List containing column-per-resample matrix for bootstrap samples and indicator matrix for OOB samples matrix
}
\description{
Bootstrap, generating in-sample and out-of-sample elements in two lists
}
\details{
Simple boostrap function
}
\examples{
# [arg_name] <- c(1, 2, 3)
# simple_bs([arg_name])

}
