% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasplit-utils.R
\name{stratifiedKfold}
\alias{stratifiedKfold}
\title{Create stratified kfold samples}
\usage{
stratifiedKfold(df, K = 5, y, balanced = TRUE)
}
\arguments{
\item{df}{a data frame}

\item{K}{number of folds, must be a integer}

\item{y}{name of target column}

\item{balanced}{makes the folds have the same proportion of classes (TRUE by default)}
}
\value{
a list with train and test elements for modeling
}
\description{
Create stratified kfold samples
}
\examples{
placeholder
}
