% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm-utils.R
\name{svm_fit_any}
\alias{svm_fit_any}
\title{Fit and Predict SVM calls on all partitions of data split}
\usage{
svm_fit_any(
  data,
  datasplit,
  svmcalls,
  metric_function,
  weight_function,
  indexes = NULL
)
}
\arguments{
\item{data}{placeholder}

\item{datasplit}{placeholder}

\item{svmcalls}{placeholder}

\item{metric_function}{placeholder}

\item{weight_function}{placeholder}

\item{indexes}{placeholder}
}
\value{
placeholder
}
\description{
Fit and Predict SVM calls on all partitions of data split
}
\examples{
placeholder
}
